Separate Form parts into different files
The various FormModel, FormConcept, FormFactory, FormContainer, etc were
getting rather difficult to mentally juggle.

This commit separates them into their own files to make it easier to
keep track.

Also, this commit adds a "maxGUI.hpp" for users to have 1 include if
they want.
