+++ Code/Examples/3-ControlGalleryExample/EntryPoint.cpp
@@ -2,23 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include <maxGUI/Button.hpp>
-#include <maxGUI/CheckBox.hpp>
-#include <maxGUI/DropDownBox.hpp>
-#include <maxGUI/EntryPoint.hpp>
-#include <maxGUI/Form.hpp>
-#include <maxGUI/Frame.hpp>
-#include <maxGUI/Label.hpp>
-#include <maxGUI/ListBox.hpp>
-#include <maxGUI/MultilineTextBox.hpp>
-#include <maxGUI/ProgressBar.hpp>
-#include <maxGUI/RadioButton.hpp>
-#include <maxGUI/Rectangle.hpp>
-#include <maxGUI/TextBox.hpp>
 #include <memory>
 #include <string>
 #include <utility>
 #include <vector>
+
+#include <maxGUI/maxGUI.hpp>
 
 class CustomButton : public maxGUI::Button
 {
