+++ Code/Examples/5-OpenGLExample/EntryPoint.cpp
@@ -2,14 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include <maxGUI/EntryPoint.hpp>
-#include <maxGUI/Form.hpp>
 #ifndef WIN32_LEAN_AND_MEAN
-#define WIN32_LEAN_AND_MEAN
+	#define WIN32_LEAN_AND_MEAN
 #endif
 #include <Windows.h>
 #include <GL/gl.h>
 #include <GL/glu.h>
+
+#include <maxGUI/maxGUI.hpp>
 
 // TODO: Provide platform-agnostic OpenGL wrappers for things like wglMakeCurrent(), etc.
