+++ Code/maxGUI/FormAllocatorConcept.cpp
@@ -1,0 +1,12 @@
+// Copyright 2022, The maxGUI Contributors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#include <maxGUI/FormAllocatorConcept.hpp>
+
+namespace maxGUI
+{
+
+	FormAllocatorConcept::~FormAllocatorConcept() noexcept = default;
+
+} // namespace maxGUI
