+++ Code/maxGUI/maxGUI.hpp
@@ -1,0 +1,27 @@
+// Copyright 2022, The maxGUI Contributors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef MAXGUI_MAXGUI_HPP
+#define MAXGUI_MAXGUI_HPP
+
+#include <maxGUI/Button.hpp>
+#include <maxGUI/CheckBox.hpp>
+#include <maxGUI/DropDownBox.hpp>
+#include <maxGUI/EntryPoint.hpp>
+#include <maxGUI/FormAllocatorConcept.hpp>
+#include <maxGUI/FormAllocatorModel.hpp>
+#include <maxGUI/FormConcept.hpp>
+#include <maxGUI/FormContainer.hpp>
+#include <maxGUI/FormFactory.hpp>
+#include <maxGUI/Frame.hpp>
+#include <maxGUI/Label.hpp>
+#include <maxGUI/ListBox.hpp>
+#include <maxGUI/MultilineTextBox.hpp>
+#include <maxGUI/ProgressBar.hpp>
+#include <maxGUI/RadioButton.hpp>
+// TODO: Move Rectangle to max
+#include <maxGUI/Rectangle.hpp>
+#include <maxGUI/TextBox.hpp>
+
+#endif // #ifndef MAXGUI_MAXGUI_HPP
