+++ Projects/VisualStudio/maxGUI/maxGUI.vcxproj
@@ -23,7 +23,9 @@
     <None Include="..\..\..\.github\ISSUE_TEMPLATE\feature_request.md" />
     <None Include="..\..\..\.github\workflows\build-and-test.yaml" />
     <None Include="..\..\..\AUTHORS" />
-    <None Include="..\..\..\Code\maxGUI\Form.inl" />
+    <None Include="..\..\..\Code\maxGUI\FormAllocatorModel.inl" />
+    <None Include="..\..\..\Code\maxGUI\FormContainer.inl" />
+    <None Include="..\..\..\Code\maxGUI\FormModel.inl" />
     <None Include="..\..\..\Docs\DeprecationSchedule.md" />
     <None Include="..\..\..\Docs\README.md" />
     <None Include="..\..\..\LICENSE" />
@@ -43,7 +45,10 @@
     <ClCompile Include="..\..\..\Code\maxGUI\ControlWithText.cpp" />
     <ClCompile Include="..\..\..\Code\maxGUI\DropDownBox.cpp" />
     <ClCompile Include="..\..\..\Code\maxGUI\EntryPoint.cpp" />
-    <ClCompile Include="..\..\..\Code\maxGUI\Form.cpp" />
+    <ClCompile Include="..\..\..\Code\maxGUI\FormAllocatorConcept.cpp" />
+    <ClCompile Include="..\..\..\Code\maxGUI\FormConcept.cpp" />
+    <ClCompile Include="..\..\..\Code\maxGUI\FormContainer.cpp" />
+    <ClCompile Include="..\..\..\Code\maxGUI\FormFactory.cpp" />
     <ClCompile Include="..\..\..\Code\maxGUI\Frame.cpp" />
     <ClCompile Include="..\..\..\Code\maxGUI\Label.cpp" />
     <ClCompile Include="..\..\..\Code\maxGUI\ListBox.cpp" />
@@ -62,10 +67,17 @@
     <ClInclude Include="..\..\..\Code\maxGUI\ControlWithText.hpp" />
     <ClInclude Include="..\..\..\Code\maxGUI\DropDownBox.hpp" />
     <ClInclude Include="..\..\..\Code\maxGUI\EntryPoint.hpp" />
-    <ClInclude Include="..\..\..\Code\maxGUI\Form.hpp" />
+    <ClInclude Include="..\..\..\Code\maxGUI\FormAllocatorConcept.hpp" />
+    <ClInclude Include="..\..\..\Code\maxGUI\FormAllocatorModel.hpp" />
+    <ClInclude Include="..\..\..\Code\maxGUI\FormConcept.hpp" />
+    <ClInclude Include="..\..\..\Code\maxGUI\FormContainer.hpp" />
+    <ClInclude Include="..\..\..\Code\maxGUI\FormFactory.hpp" />
+    <ClInclude Include="..\..\..\Code\maxGUI\FormModel.hpp" />
+    <ClInclude Include="..\..\..\Code\maxGUI\FormStyles.hpp" />
     <ClInclude Include="..\..\..\Code\maxGUI\Frame.hpp" />
     <ClInclude Include="..\..\..\Code\maxGUI\Label.hpp" />
     <ClInclude Include="..\..\..\Code\maxGUI\ListBox.hpp" />
+    <ClInclude Include="..\..\..\Code\maxGUI\maxGUI.hpp" />
     <ClInclude Include="..\..\..\Code\maxGUI\MultilineTextBox.hpp" />
     <ClInclude Include="..\..\..\Code\maxGUI\ProgressBar.hpp" />
     <ClInclude Include="..\..\..\Code\maxGUI\RadioButton.hpp" />
